Added a new button 'Reset' and changed logic to incorporate the change. Additionally, made some housekeeping changes.

+++ app/src/main/java/com/example/myapplication/SoundMeter.java
@@ -56,7 +56,7 @@ public class SoundMeter {
         float numCycles = numCrossing/2;
         float frequency = numCycles/numSecondsRecorded;
 
-        return (double)frequency;
+        return frequency;
     }
 
 }
